+++ meerk40t/gui/wxmmain.py
@@ -10,6 +10,7 @@ from PIL import Image
 from wx import aui
 
 from meerk40t.core.exceptions import BadFileError
+from meerk40t.gui.gui_mixins import FormatPainter, Warnings
 from meerk40t.gui.statusbarwidgets.defaultoperations import DefaultOperationWidget
 from meerk40t.gui.statusbarwidgets.infowidget import (
     BurnProgressPanel,
@@ -124,169 +125,6 @@ from .mwindow import MWindow
 
 _ = wx.GetTranslation
 
-INACTIVE = "inactive"
-WAITING = "waiting"
-PASTING = "pasting"
-
-
-class FormatPainter:
-    def __init__(self, context, button, identifier, *args, **kwds):
-        self.context = context
-        # Path to button
-        self.button = button
-        self.identifier = identifier
-        # The node to use
-        self.template = None
-        # List of tuples with (attribute_name, generic)
-        self.possible_attributes = (
-            # Standard line and fill attributes
-            ("stroke", True),
-            ("stroke_width", True),
-            ("stroke_scale", True),
-            ("fill", True),
-            ("linecap", True),
-            ("linejoin", True),
-            ("fillrule", True),
-            # Image attributes
-            ("dpi", False),
-            ("operations", False),
-            ("invert", False),
-            ("dither", False),
-            ("dither_type", False),
-            ("red", False),
-            ("green", False),
-            ("blue", False),
-            ("lightness", False),
-            # Text attributes
-            ("mkfont", True),
-            ("mkfontsize", True),
-            ("font_style", False),
-            ("font_variant", False),
-            ("font_weight", False),
-            ("font_stretch", False),
-            ("font_size", False),
-            ("line_height", False),
-            ("font_family", False),
-            # Hatches
-            ("hatch_distance", False),
-            ("hatch_angle", False),
-            ("hatch_angle_delta", False),
-            ("hatch_type", False),
-            # Wobbles
-            ("wobble_radius", False),
-            ("wobble_interval", False),
-            ("wobble_speed", False),
-            ("wobble_type", False),
-        )
-        self.path_update_needed = (
-            "mkfont",
-            "mkfontsize",
-        )
-        # State-Machine
-        self._state = None
-        self.state = INACTIVE
-
-    @property
-    def state(self):
-        return self._state
-
-    @state.setter
-    def state(self, value):
-        if value not in (INACTIVE, WAITING, PASTING):
-            value = INACTIVE
-        if value == INACTIVE:
-            self.template = None
-        elif value == WAITING:
-            node = self.context.elements.first_emphasized
-            if node is not None and node.type not in ("file", "group"):
-                self.template = node
-                value = PASTING
-        elif value == PASTING:
-            pass
-        self._state = value
-        toggle = bool(value != INACTIVE)
-        self.context.signal(self.identifier, toggle)
-
-    def on_emphasis(self, *args):
-        if self.state == INACTIVE:
-            return
-        elif self.state == WAITING:
-            node = self.context.elements.first_emphasized
-            if node is None:
-                return
-            if node.type in ("file", "group"):
-                return
-            self.template = node
-            self.state = PASTING
-            return
-        elif self.state == PASTING:
-            try:
-                id = self.template.id
-            except (RuntimeError, AttributeError):
-                # No longer existing or invalid?
-                self.state = INACTIVE
-            nodes_changed = []
-            nodes_classify = []
-            nodes_images = []
-            for node in self.context.elements.elems(emphasized=True):
-                if node is self.template:
-                    continue
-                flag_changed = False
-                flag_classify = False
-                flag_pathupdate = False
-                for entry in self.possible_attributes:
-                    attr = entry[0]
-                    generic = entry[1]
-                    if not generic and node.type != self.template.type:
-                        continue
-                    if hasattr(self.template, attr) and hasattr(node, attr):
-                        value = getattr(self.template, attr, None)
-                        if isinstance(value, (list, tuple)):
-                            value = copy(value)
-                        try:
-                            setattr(node, attr, value)
-                            flag_changed = True
-                            if attr in ("stroke", "fill"):
-                                flag_classify = True
-                            if attr in self.path_update_needed:
-                                flag_pathupdate = True
-                        except ValueError:
-                            continue
-                if flag_changed:
-                    nodes_changed.append(node)
-                    if node.type == "elem image":
-                        nodes_images.append(node)
-                if flag_pathupdate:
-                    if hasattr(node, "mktext"):
-                        newtext = self.context.elements.wordlist_translate(
-                            node.mktext, elemnode=node, increment=False
-                        )
-                        oldtext = getattr(node, "_translated_text", "")
-                        if newtext != oldtext:
-                            node._translated_text = newtext
-                        kernel = self.context.kernel
-                        for property_op in kernel.lookup_all("path_updater/.*"):
-                            property_op(kernel.root, node)
-                        if hasattr(node, "_cache"):
-                            node._cache = None
-
-                if flag_classify:
-                    nodes_classify.append(node)
-            if len(nodes_changed) > 0:
-                for node in nodes_images:
-                    node.update(None)
-                if len(nodes_classify) > 0 and self.context.elements.classify_new:
-                    self.context.elements.classify(nodes_classify)
-                self.context.signal("element_property_update", nodes_changed)
-                self.context.signal("refresh_scene", "Scene")
-
-    def on_click(self, *args):
-        # print(f"On_click called, state was : {self.state}")
-        if self.state == INACTIVE:
-            self.state = WAITING
-        else:
-            self.state = INACTIVE
-
 
 class MeerK40t(MWindow):
     """MeerK40t main window"""
@@ -397,12 +235,6 @@ class MeerK40t(MWindow):
         self.update_check_at_startup()
         self.tips_at_startup()
         self.parametric_info = None
-        self._concerns = list()
-        self._device_acceleration_info = dict()
-
-    @property
-    def concerns(self):
-        return "\n".join(self._concerns)
 
     def tips_at_startup(self):
         self.context.setting(bool, "show_tips", True)
@@ -1326,48 +1158,9 @@ class MeerK40t(MWindow):
             set_icon_appearance(1.0, 0)
 
         self.format_painter = FormatPainter(
-            self.context, "button/basicediting/Paint", "editpaint"
+            self.context, "button/extended_tools/Paint", "editpaint"
         )
-        self.context.kernel.register(
-            "button/basicediting/Paint",
-            {
-                "label": _("Paint format"),
-                "icon": icon_paint_brush,
-                "tip": _(
-                    "First select your template, then every subsequent selection will apply the templates properties to the selected elements"
-                ),
-                "help": "basicediting",
-                "action": self.format_painter.on_click,
-                "identifier": "editpaint",
-                "toggle": {
-                    "label": _("Stop"),
-                    "help": "basicediting",
-                    "action": self.format_painter.on_click,
-                    "icon": icon_paint_brush_green,
-                    "signal": "editpaint",
-                    "tip": _("Click again to disable the paint mode"),
-                },
-            },
-        )
-
-        def show_concerns(*args):
-            if len(self._concerns):
-                wx.MessageBox(
-                    self.concerns, _("Warning"), style=wx.OK | wx.ICON_WARNING
-                )
-
-        self.context.kernel.register(
-            "button/jobstart/Warning",
-            {
-                "label": _("Warning"),
-                "icon": icon_warning,
-                "rule_visible": lambda d: len(self._concerns) > 0,
-                "action": show_concerns,
-                "tip": _("There are issues with your project"),
-                "size": STD_ICON_SIZE,
-                "priority": 2,
-            },
-        )
+        self.warning_routine = Warnings(self.context, "button/jobstart/Warning")
 
     def open_property_window_for_node(self, node):
         """
@@ -4061,6 +3854,12 @@ class MeerK40t(MWindow):
             #     m.Enable(False)
         self.main_menubar.Append(wxglade_tmp_menu, _("Languages"))
 
+    @signal_listener("warn_state_update")
+    @signal_listener("updateop_tree")
+    @signal_listener("tool_modified")
+    def warning_indicator(self, *args):
+        self.warning_routine.warning_indicator()
+
     @signal_listener("restart")
     def on_restart_required(self, *args):
         self.context.kernel.register(
@@ -4073,192 +3872,6 @@ class MeerK40t(MWindow):
                 "size": STD_ICON_SIZE,
             },
         )
-
-    @signal_listener("warn_state_update")
-    @signal_listener("updateop_tree")
-    @signal_listener("tool_modified")
-    def warning_indicator(self, *args):
-        def has_ambitious_operations(maxspeed, optypes):
-            for op in self.context.elements.ops():
-                if (
-                    hasattr(op, "output")
-                    and hasattr(op, "speed")
-                    and op.output
-                    and op.type in optypes
-                    and len(op.children) > 0
-                ):
-                    # Is a warning defined?
-                    checker = f"dangerlevel_{op.type.replace(' ', '_')}"
-                    danger = False
-                    if hasattr(self.context.device, checker):
-                        maxspeed_minpower = getattr(self.context.device, checker)
-                        if (
-                            isinstance(maxspeed_minpower, (tuple, list))
-                            and len(maxspeed_minpower) == 8
-                        ):
-                            # minpower, maxposer, minspeed, maxspeed
-                            # print ("Yes: ", checker, maxspeed_minpower)
-                            danger = False
-                            if hasattr(op, "power"):
-                                value = op.power
-                                if (
-                                    maxspeed_minpower[0]
-                                    and value < maxspeed_minpower[1]
-                                ):
-                                    danger = True
-                                if (
-                                    maxspeed_minpower[2]
-                                    and value > maxspeed_minpower[3]
-                                ):
-                                    danger = True
-                            if hasattr(op, "speed"):
-                                value = op.speed
-                                if (
-                                    maxspeed_minpower[4]
-                                    and value < maxspeed_minpower[5]
-                                ):
-                                    danger = True
-                                if (
-                                    maxspeed_minpower[6]
-                                    and value > maxspeed_minpower[7]
-                                ):
-                                    danger = True
-                    if danger:
-                        return True
-                    # Is a generic maximum speed defined?
-                    if maxspeed is not None and op.speed >= maxspeed:
-                        return True
-
-            return False
-
-        def has_objects_outside():
-            wd = self.context.space.display.width
-            ht = self.context.space.display.height
-            for op in self.context.elements.ops():
-                if hasattr(op, "output") and op.output:
-                    for refnode in op.children:
-                        node = refnode.node
-                        bb = getattr(node, "paint_bounds", None)
-                        if bb is None:
-                            bb = getattr(node, "bounds", None)
-                        if bb is None:
-                            continue
-                        if bb[2] > wd or bb[0] < 0 or bb[3] > ht or bb[1] < 0:
-                            return True
-            return False
-
-        def has_close_to_edge_rasters():
-            wd = self.context.space.display.width
-            ht = self.context.space.display.height
-            additional_info = ""
-            devname = self.context.device.name
-            if devname not in self._device_acceleration_info:
-                acceleration = False
-                device_dict = dir(self.context.device)
-                for d in device_dict:
-                    if "acceler" in d.lower():
-                        acceleration = True
-                        break
-                if not acceleration and hasattr(self.context.device, "driver"):
-                    device_dict = dir(self.context.device.driver)
-                    for d in device_dict:
-                        if "acceler" in d.lower():
-                            acceleration = True
-                            break
-                self._device_acceleration_info[devname] = acceleration
-            acceleration = self._device_acceleration_info[devname]
-
-            for op in self.context.elements.ops():
-                if (
-                    hasattr(op, "output")
-                    and op.output
-                    and op.type in ("op raster", "op image")
-                ):
-                    dx = 0
-                    dy = 0
-                    if hasattr(op, "overscan") and op.overscan is not None:
-                        try:
-                            ov = float(op.overscan)
-                        except ValueError:
-                            ov = 0
-                        dx += ov
-                        dy += ov
-                    if acceleration:
-                        # Acceleration / deacceleration plays a role.
-                        if hasattr(op, "speed") and op.speed:
-                            a = 500  # arbitrary 500 mm/sec²
-                            dt = op.speed / a
-                            ds = 0.5 * a * dt * dt
-                            dx += ds * UNITS_PER_MM
-
-                    for refnode in op.children:
-                        node = refnode.node
-                        bb = getattr(node, "paint_bounds", None)
-                        if bb is None:
-                            bb = getattr(node, "bounds", None)
-                        if bb is None:
-                            continue
-                        if bb[2] > wd or bb[0] < 0 or bb[3] > ht or bb[1] < 0:
-                            # Even though is bad, that's not what we are looking for
-                            continue
-                        flag = False
-                        if bb[2] + dx > wd or bb[0] - dx < 0:
-                            if additional_info:
-                                additional_info += ", "
-                            additional_info += f"x > {Length(dx, digits=1).length_mm}"
-                            flag = True
-                        if bb[3] + dy > ht or bb[1] - dy < 0:
-                            if additional_info:
-                                additional_info += ", "
-                            additional_info += f"y > {Length(dy, digits=1).length_mm}"
-                            return True, additional_info
-                        if flag:
-                            return True, additional_info
-
-            return False, ""
-
-        self._concerns.clear()
-        max_speed = getattr(self.context.device, "max_vector_speed", None)
-        if has_ambitious_operations(max_speed, ("op cut", "op engrave")):
-            self._concerns.append(
-                _("- Vector operations are too fast.")
-                + "\n  "
-                + _("Could lead to erratic stepper behaviour and incomplete burns.")
-            )
-        max_speed = getattr(self.context.device, "max_raster_speed", None)
-        if has_ambitious_operations(max_speed, ("op raster", "op image")):
-            self._concerns.append(
-                _("- Raster operations are too fast.")
-                + "\n  "
-                + _("Could lead to erratic stepper behaviour and incomplete burns.")
-            )
-        if has_objects_outside():
-            self._concerns.append(
-                _("- Elements are lying outside the burnable area.")
-                + "\n  "
-                + _("Could lead to the laserhead bumping into the rails.")
-            )
-        flag, info = has_close_to_edge_rasters()
-        if flag:
-            self._concerns.append(
-                _("- Raster operations get very close to the edge.")
-                + "\n  "
-                + _("Could lead to the laserhead bumping into the rails.")
-                + "\n  "
-                + info
-            )
-
-        non_assigned, non_burn = self.context.elements.have_unburnable_elements()
-        if non_assigned:
-            self._concerns.append(
-                _("- Elements aren't assigned to an operation and will not be burnt")
-            )
-        if non_burn:
-            self._concerns.append(
-                _(
-                    "- Some operations containing elements aren't active, so some elements will not be burnt"
-                )
-            )
 
     @signal_listener("file;loaded")
     @signal_listener("file;saved")
